Rewrote chart as ES6 module
