+++ .eslintrc
@@ -16,5 +16,7 @@
         "prefer-spread": 0,
         "no-shadow": 0,
         "no-underscore-dangle": 0,
+        "import/prefer-default-export": 0,
+        "one-var-declaration-per-line": 0,
     }
 }
