+++ es6_modules/chart/api.color.js
@@ -1,0 +1,26 @@
+import {
+    CLASS,
+    isValue,
+    isFunction,
+    isString,
+    isUndefined,
+    isDefined,
+    ceil10,
+    asHalfPixel,
+    diffDomain,
+    isEmpty,
+    notEmpty,
+    getOption,
+    hasValue,
+    sanitise,
+    getPathBox,
+    ChartInternal
+} from '../chartinternal.js';
+
+// TODO: fix
+const color = function (id) {
+    const $$ = this.internal;
+    return $$.color(id); // more patterns
+};
+
+export { color };
