+++ es6_modules/chart/api.group.js
@@ -1,0 +1,28 @@
+import {
+    CLASS,
+    isValue,
+    isFunction,
+    isString,
+    isUndefined,
+    isDefined,
+    ceil10,
+    asHalfPixel,
+    diffDomain,
+    isEmpty,
+    notEmpty,
+    getOption,
+    hasValue,
+    sanitise,
+    getPathBox,
+    ChartInternal
+} from '../chartinternal.js';
+
+const groups = function (groups) {
+    let $$ = this.internal, config = $$.config;
+    if (isUndefined(groups)) { return config.data_groups; }
+    config.data_groups = groups;
+    $$.redraw();
+    return config.data_groups;
+};
+
+export { groups };
