+++ es6_modules/chart/api.legend.js
@@ -1,0 +1,34 @@
+import {
+    CLASS,
+    isValue,
+    isFunction,
+    isString,
+    isUndefined,
+    isDefined,
+    ceil10,
+    asHalfPixel,
+    diffDomain,
+    isEmpty,
+    notEmpty,
+    getOption,
+    hasValue,
+    sanitise,
+    getPathBox,
+    ChartInternal
+} from '../chartinternal.js';
+
+const legend = function () {};
+
+legend.show = function (targetIds) {
+    const $$ = this.internal;
+    $$.showLegend($$.mapToTargetIds(targetIds));
+    $$.updateAndRedraw({ withLegend: true });
+};
+
+legend.hide = function (targetIds) {
+    const $$ = this.internal;
+    $$.hideLegend($$.mapToTargetIds(targetIds));
+    $$.updateAndRedraw({ withLegend: true });
+};
+
+export { legend };
