+++ es6_modules/chart/api.transform.js
@@ -1,0 +1,26 @@
+import {
+    CLASS,
+    isValue,
+    isFunction,
+    isString,
+    isUndefined,
+    isDefined,
+    ceil10,
+    asHalfPixel,
+    diffDomain,
+    isEmpty,
+    notEmpty,
+    getOption,
+    hasValue,
+    sanitise,
+    getPathBox,
+    ChartInternal
+} from '../chartinternal.js';
+
+const transform = function (type, targetIds) {
+    let $$ = this.internal,
+        options = ['pie', 'donut'].indexOf(type) >= 0 ? { withTransform: true } : null;
+    $$.transformTo(targetIds, type, options);
+};
+
+export { transform };
