+++ es6_modules/chart/api.x.js
@@ -1,0 +1,38 @@
+import {
+    CLASS,
+    isValue,
+    isFunction,
+    isString,
+    isUndefined,
+    isDefined,
+    ceil10,
+    asHalfPixel,
+    diffDomain,
+    isEmpty,
+    notEmpty,
+    getOption,
+    hasValue,
+    sanitise,
+    getPathBox,
+    ChartInternal
+} from '../chartinternal.js';
+
+const x = function (x) {
+    const $$ = this.internal;
+    if (arguments.length) {
+        $$.updateTargetX($$.data.targets, x);
+        $$.redraw({ withUpdateOrgXDomain: true, withUpdateXDomain: true });
+    }
+    return $$.data.xs;
+};
+
+const xs = function (xs) {
+    const $$ = this.internal;
+    if (arguments.length) {
+        $$.updateTargetXs($$.data.targets, xs);
+        $$.redraw({ withUpdateOrgXDomain: true, withUpdateXDomain: true });
+    }
+    return $$.data.xs;
+};
+
+export { x, xs };
